refactor(seat): SeatHoldService 트랜잭션 로직을 별도 Bean으로 분리하여 락-커밋 순서 보장
기존에는 SeatHoldService.createOrRefreshHold()에 @Transactional이 걸려 있어 finally 블록에서 unlock이 먼저 실행되고, Spring AOP 프록시가 이후에 커밋하는 구조였음.
이로 인해 다른 스레드가 커밋되지 않은 데이터를 읽을 수 있는 위험이 존재.

- SeatHoldTransactionalService를 별도 빈으로 분리하여 트랜잭션 커밋이 완료된 후 finally에서 unlock되도록 순서를 보장함.
- 추가로 releaseUserActiveHolds에서 delete 후 flush()를 호출하여 JPA flush 순서(INSERT → UPDATE → DELETE)로 인한 unique constraint violation을 방지함
